requirements.txt
fastapi==0.98.0
uvicorn==0.23.0
redis==4.5.5
python-dotenv==1.0.0
email-validator==1.3.1
